Break out `formatQuery` tests by format
