+++ packages/react-querybuilder/src/utils/formatQuery/formatQuery.test.ts
@@ -11,529 +11,20 @@ import type {
 } from '../../types/index.noReact';
 import { convertToIC } from '../convertQuery';
 import { prepareRuleGroup } from '../prepareQueryObjects';
-import { add } from '../queryTools';
-import { defaultRuleProcessorCEL } from './defaultRuleProcessorCEL';
-import { defaultRuleProcessorElasticSearch } from './defaultRuleProcessorElasticSearch';
-import { defaultRuleProcessorJsonLogic } from './defaultRuleProcessorJsonLogic';
-import { defaultRuleProcessorMongoDB } from './defaultRuleProcessorMongoDB';
-import { defaultRuleProcessorSpEL } from './defaultRuleProcessorSpEL';
 import { defaultRuleProcessorSQL } from './defaultRuleProcessorSQL';
 import { formatQuery } from './formatQuery';
 import {
-  defaultCELValueProcessor,
-  defaultMongoDBValueProcessor,
-  defaultSpELValueProcessor,
-  defaultValueProcessor,
-  defaultValueProcessorByRule,
-} from './index';
-import { jsonLogicAdditionalOperators, quoteFieldNamesWithArray } from './utils';
-
-const query: RuleGroupType = {
-  id: 'g-root',
-  rules: [
-    { field: defaultFieldPlaceholder, operator: defaultOperatorPlaceholder, value: 'Placeholder' },
-    { field: defaultFieldPlaceholder, operator: '=', value: 'Placeholder' },
-    { field: 'firstName', operator: defaultOperatorPlaceholder, value: 'Placeholder' },
-    { field: 'firstName', operator: 'null', value: '' },
-    { field: 'lastName', operator: 'notNull', value: '' },
-    { field: 'firstName', operator: 'in', value: 'Test,This' },
-    { field: 'lastName', operator: 'notIn', value: 'Test,This' },
-    { field: 'firstName', operator: 'in', value: false },
-    { field: 'firstName', operator: 'between', value: 'Test,This' },
-    { field: 'firstName', operator: 'between', value: ['Test', 'This'] },
-    { field: 'lastName', operator: 'notBetween', value: 'Test,This' },
-    { field: 'firstName', operator: 'between', value: 'MissingComma' },
-    { field: 'age', operator: 'between', value: '12,14' },
-    { field: 'firstName', operator: 'between', value: 'OnlyFirstElement,' },
-    { field: 'firstName', operator: 'between', value: ',OnlySecondElement' },
-    { field: 'age', operator: '=', value: '26' },
-    { field: 'isMusician', operator: '=', value: true },
-    { field: 'isLucky', operator: '=', value: false },
-    {
-      id: 'g-sub1',
-      combinator: 'or',
-      rules: [
-        { field: 'gender', operator: '=', value: 'M' },
-        { field: 'job', operator: '!=', value: 'Programmer' },
-        { field: 'email', operator: 'contains', value: '@' },
-      ],
-      not: true,
-    },
-    {
-      id: 'g-sub2',
-      combinator: 'or',
-      rules: [
-        { field: 'lastName', operator: 'doesNotContain', value: 'ab' },
-        { field: 'job', operator: 'beginsWith', value: 'Prog' },
-        { field: 'email', operator: 'endsWith', value: 'com' },
-        { field: 'job', operator: 'doesNotBeginWith', value: 'Man' },
-        { field: 'email', operator: 'doesNotEndWith', value: 'fr' },
-      ],
-      not: false,
-    },
-  ],
-  combinator: 'and',
-  not: false,
-};
-const mongoQuery: RuleGroupType = {
-  id: 'g-root',
-  combinator: 'and',
-  rules: [
-    { field: '~', operator: '~', value: 'Placeholder' },
-    { field: '~', operator: '=', value: 'Placeholder' },
-    { field: 'firstName', operator: '~', value: 'Placeholder' },
-    { field: 'invalid', value: '', operator: 'invalid' },
-    { field: 'firstName', value: '', operator: 'null' },
-    { field: 'lastName', value: '', operator: 'notNull' },
-    { field: 'firstName', value: 'Test,This', operator: 'in' },
-    { field: 'lastName', value: 'Test,This', operator: 'notIn' },
-    { field: 'firstName', value: false, operator: 'in' },
-    { field: 'firstName', value: 'Test,This', operator: 'between' },
-    { field: 'firstName', value: ['Test', 'This'], operator: 'between' },
-    { field: 'lastName', value: 'Test,This', operator: 'notBetween' },
-    { field: 'firstName', value: '', operator: 'between' },
-    { field: 'firstName', value: false, operator: 'between' },
-    { field: 'age', value: '12,14', operator: 'between' },
-    { field: 'age', value: '26', operator: '=' },
-    { field: 'isMusician', value: true, operator: '=' },
-    { field: 'isLucky', value: false, operator: '=' },
-    { field: 'email', value: '@', operator: 'contains' },
-    { field: 'email', value: 'ab', operator: 'beginsWith' },
-    { field: 'email', value: 'com', operator: 'endsWith' },
-    { field: 'hello', value: 'com', operator: 'doesNotContain' },
-    { field: 'job', value: 'Man', operator: 'doesNotBeginWith' },
-    { field: 'job', value: 'ger', operator: 'doesNotEndWith' },
-    {
-      id: 'g-sub1',
-      combinator: 'or',
-      rules: [
-        { field: 'job', value: 'Sales Executive', operator: '=' },
-        { field: 'job', value: [], operator: 'in' },
-        { field: 'job', value: ['just one value'], operator: 'between' },
-      ],
-      not: false,
-    },
-  ],
-  not: false,
-};
-const queryWithValueSourceField: RuleGroupType = {
-  combinator: 'and',
-  rules: [
-    { field: 'firstName', operator: 'null', value: '', valueSource: 'field' },
-    { field: 'lastName', operator: 'notNull', value: '', valueSource: 'field' },
-    { field: 'firstName', operator: 'in', value: 'middleName,lastName', valueSource: 'field' },
-    { field: 'lastName', operator: 'notIn', value: 'middleName,lastName', valueSource: 'field' },
-    { field: 'firstName', operator: 'in', value: false, valueSource: 'field' },
-    { field: 'firstName', operator: 'between', value: 'middleName,lastName', valueSource: 'field' },
-    {
-      field: 'firstName',
-      operator: 'between',
-      value: ['middleName', 'lastName'],
-      valueSource: 'field',
-    },
-    {
-      field: 'lastName',
-      operator: 'notBetween',
-      value: 'middleName,lastName',
-      valueSource: 'field',
-    },
-    { field: 'firstName', operator: 'between', value: 'MissingComma', valueSource: 'field' },
-    { field: 'age', operator: '=', value: 'iq', valueSource: 'field' },
-    { field: 'isMusician', operator: '=', value: 'isCreative', valueSource: 'field' },
-    {
-      id: 'g-sub1',
-      combinator: 'or',
-      rules: [
-        { field: 'gender', operator: '=', value: 'someLetter', valueSource: 'field' },
-        { field: 'job', operator: '!=', value: 'isBetweenJobs', valueSource: 'field' },
-        { field: 'email', operator: 'contains', value: 'atSign', valueSource: 'field' },
-      ],
-      not: true,
-    },
-    {
-      id: 'g-sub2',
-      combinator: 'or',
-      rules: [
-        { field: 'lastName', operator: 'doesNotContain', value: 'firstName', valueSource: 'field' },
-        { field: 'job', operator: 'beginsWith', value: 'jobPrefix', valueSource: 'field' },
-        { field: 'email', operator: 'endsWith', value: 'dotCom', valueSource: 'field' },
-        { field: 'job', operator: 'doesNotBeginWith', value: 'hasNoJob', valueSource: 'field' },
-        { field: 'email', operator: 'doesNotEndWith', value: 'isInvalid', valueSource: 'field' },
-      ],
-      not: false,
-    },
-  ],
-};
-const mongoQueryWithValueSourceField: RuleGroupType = {
-  id: 'g-root',
-  combinator: 'and',
-  rules: [
-    { field: 'invalid', operator: 'invalid', value: '', valueSource: 'field' },
-    { field: 'firstName', operator: 'null', value: '', valueSource: 'field' },
-    { field: 'lastName', operator: 'notNull', value: '', valueSource: 'field' },
-    { field: 'firstName', operator: 'in', value: 'middleName,lastName', valueSource: 'field' },
-    { field: 'lastName', operator: 'notIn', value: 'middleName,lastName', valueSource: 'field' },
-    { field: 'firstName', operator: 'in', value: false, valueSource: 'field' },
-    { field: 'firstName', operator: 'between', value: 'middleName,lastName', valueSource: 'field' },
-    {
-      field: 'firstName',
-      operator: 'between',
-      value: ['middleName', 'lastName'],
-      valueSource: 'field',
-    },
-    {
-      field: 'lastName',
-      operator: 'notBetween',
-      value: 'middleName,lastName',
-      valueSource: 'field',
-    },
-    { field: 'firstName', operator: 'between', value: '', valueSource: 'field' },
-    { field: 'firstName', operator: 'between', value: false, valueSource: 'field' },
-    { field: 'age', operator: '=', value: 'iq', valueSource: 'field' },
-    { field: 'isMusician', operator: '=', value: 'isCreative', valueSource: 'field' },
-    { field: 'email', operator: 'contains', value: 'atSign', valueSource: 'field' },
-    { field: 'email', operator: 'beginsWith', value: 'name', valueSource: 'field' },
-    { field: 'email', operator: 'endsWith', value: 'dotCom', valueSource: 'field' },
-    { field: 'hello', operator: 'doesNotContain', value: 'dotCom', valueSource: 'field' },
-    { field: 'job', operator: 'doesNotBeginWith', value: 'noJob', valueSource: 'field' },
-    { field: 'job', operator: 'doesNotEndWith', value: 'noJob', valueSource: 'field' },
-    {
-      id: 'g-sub1',
-      combinator: 'or',
-      rules: [{ field: 'job', operator: '=', value: 'executiveJobName', valueSource: 'field' }],
-      not: false,
-    },
-  ],
-  not: false,
-};
-
-const sqlString =
-  "(firstName is null and lastName is not null and firstName in ('Test', 'This') and lastName not in ('Test', 'This') and firstName between 'Test' and 'This' and firstName between 'Test' and 'This' and lastName not between 'Test' and 'This' and age between '12' and '14' and age = '26' and isMusician = TRUE and isLucky = FALSE and NOT (gender = 'M' or job != 'Programmer' or email like '%@%') and (lastName not like '%ab%' or job like 'Prog%' or email like '%com' or job not like 'Man%' or email not like '%fr'))";
-const sqlStringForValueSourceField =
-  "(firstName is null and lastName is not null and firstName in (middleName, lastName) and lastName not in (middleName, lastName) and firstName between middleName and lastName and firstName between middleName and lastName and lastName not between middleName and lastName and age = iq and isMusician = isCreative and NOT (gender = someLetter or job != isBetweenJobs or email like '%' || atSign || '%') and (lastName not like '%' || firstName || '%' or job like jobPrefix || '%' or email like '%' || dotCom or job not like hasNoJob || '%' or email not like '%' || isInvalid))";
-const parameterizedSQLString =
-  '(firstName is null and lastName is not null and firstName in (?, ?) and lastName not in (?, ?) and firstName between ? and ? and firstName between ? and ? and lastName not between ? and ? and age between ? and ? and age = ? and isMusician = ? and isLucky = ? and NOT (gender = ? or job != ? or email like ?) and (lastName not like ? or job like ? or email like ? or job not like ? or email not like ?))';
-const parameterizedNamedSQLString =
-  '(firstName is null and lastName is not null and firstName in (:firstName_1, :firstName_2) and lastName not in (:lastName_1, :lastName_2) and firstName between :firstName_3 and :firstName_4 and firstName between :firstName_5 and :firstName_6 and lastName not between :lastName_3 and :lastName_4 and age between :age_1 and :age_2 and age = :age_3 and isMusician = :isMusician_1 and isLucky = :isLucky_1 and NOT (gender = :gender_1 or job != :job_1 or email like :email_1) and (lastName not like :lastName_5 or job like :job_2 or email like :email_2 or job not like :job_3 or email not like :email_3))';
-const params = [
-  'Test',
-  'This',
-  'Test',
-  'This',
-  'Test',
-  'This',
-  'Test',
-  'This',
-  'Test',
-  'This',
-  '12',
-  '14',
-  '26',
-  true,
-  false,
-  'M',
-  'Programmer',
-  '%@%',
-  '%ab%',
-  'Prog%',
-  '%com',
-  'Man%',
-  '%fr',
-];
-const params_named = {
-  firstName_1: 'Test',
-  firstName_2: 'This',
-  lastName_1: 'Test',
-  lastName_2: 'This',
-  firstName_3: 'Test',
-  firstName_4: 'This',
-  firstName_5: 'Test',
-  firstName_6: 'This',
-  lastName_3: 'Test',
-  lastName_4: 'This',
-  age_1: '12',
-  age_2: '14',
-  age_3: '26',
-  isMusician_1: true,
-  isLucky_1: false,
-  gender_1: 'M',
-  job_1: 'Programmer',
-  email_1: '%@%',
-  lastName_5: '%ab%',
-  job_2: 'Prog%',
-  email_2: '%com',
-  job_3: 'Man%',
-  email_3: '%fr',
-};
-const mongoQueryExpectation = {
-  $and: [
-    { firstName: null },
-    { lastName: { $ne: null } },
-    { firstName: { $in: ['Test', 'This'] } },
-    { lastName: { $nin: ['Test', 'This'] } },
-    { firstName: { $gte: 'Test', $lte: 'This' } },
-    { firstName: { $gte: 'Test', $lte: 'This' } },
-    { $or: [{ lastName: { $lt: 'Test' } }, { lastName: { $gt: 'This' } }] },
-    { age: { $gte: 12, $lte: 14 } },
-    { age: '26' },
-    { isMusician: true },
-    { isLucky: false },
-    { email: { $regex: '@' } },
-    { email: { $regex: '^ab' } },
-    { email: { $regex: 'com$' } },
-    { hello: { $not: { $regex: 'com' } } },
-    { job: { $not: { $regex: '^Man' } } },
-    { job: { $not: { $regex: 'ger$' } } },
-    { job: 'Sales Executive' },
-  ],
-};
-const mongoQueryExpectationForValueSourceField = {
-  $and: [
-    { firstName: null },
-    { lastName: { $ne: null } },
-    { $where: '[this.middleName,this.lastName].includes(this.firstName)' },
-    { $where: '![this.middleName,this.lastName].includes(this.lastName)' },
-    {
-      $and: [
-        { $expr: { $gte: ['$firstName', '$middleName'] } },
-        { $expr: { $lte: ['$firstName', '$lastName'] } },
-      ],
-    },
-    {
-      $and: [
-        { $expr: { $gte: ['$firstName', '$middleName'] } },
-        { $expr: { $lte: ['$firstName', '$lastName'] } },
-      ],
-    },
-    {
-      $or: [
-        { $expr: { $lt: ['$lastName', '$middleName'] } },
-        { $expr: { $gt: ['$lastName', '$lastName'] } },
-      ],
-    },
-    { $expr: { $eq: ['$age', '$iq'] } },
-    { $expr: { $eq: ['$isMusician', '$isCreative'] } },
-    { $where: 'this.email.includes(this.atSign)' },
-    { $where: 'this.email.startsWith(this.name)' },
-    { $where: 'this.email.endsWith(this.dotCom)' },
-    { $where: '!this.hello.includes(this.dotCom)' },
-    { $where: '!this.job.startsWith(this.noJob)' },
-    { $where: '!this.job.endsWith(this.noJob)' },
-    { $expr: { $eq: ['$job', '$executiveJobName'] } },
-  ],
-};
-const celString =
-  'firstName == null && lastName != null && firstName in ["Test", "This"] && !(lastName in ["Test", "This"]) && (firstName >= "Test" && firstName <= "This") && (firstName >= "Test" && firstName <= "This") && (lastName < "Test" || lastName > "This") && (age >= 12 && age <= 14) && age == "26" && isMusician == true && isLucky == false && !(gender == "M" || job != "Programmer" || email.contains("@")) && (!lastName.contains("ab") || job.startsWith("Prog") || email.endsWith("com") || !job.startsWith("Man") || !email.endsWith("fr"))';
-const celStringForValueSourceField =
-  'firstName == null && lastName != null && firstName in [middleName, lastName] && !(lastName in [middleName, lastName]) && (firstName >= middleName && firstName <= lastName) && (firstName >= middleName && firstName <= lastName) && (lastName < middleName || lastName > lastName) && age == iq && isMusician == isCreative && !(gender == someLetter || job != isBetweenJobs || email.contains(atSign)) && (!lastName.contains(firstName) || job.startsWith(jobPrefix) || email.endsWith(dotCom) || !job.startsWith(hasNoJob) || !email.endsWith(isInvalid))';
-const spelString =
-  "firstName == null and lastName != null and (firstName == 'Test' or firstName == 'This') and !(lastName == 'Test' or lastName == 'This') and (firstName >= 'Test' and firstName <= 'This') and (firstName >= 'Test' and firstName <= 'This') and (lastName < 'Test' or lastName > 'This') and (age >= 12 and age <= 14) and age == '26' and isMusician == true and isLucky == false and !(gender == 'M' or job != 'Programmer' or email matches '@') and (!(lastName matches 'ab') or job matches '^Prog' or email matches 'com$' or !(job matches '^Man') or !(email matches 'fr$'))";
-const spelStringForValueSourceField =
-  "firstName == null and lastName != null and (firstName == middleName or firstName == lastName) and !(lastName == middleName or lastName == lastName) and (firstName >= middleName and firstName <= lastName) and (firstName >= middleName and firstName <= lastName) and (lastName < middleName or lastName > lastName) and age == iq and isMusician == isCreative and !(gender == someLetter or job != isBetweenJobs or email matches atSign) and (!(lastName matches firstName) or job matches '^'.concat(jobPrefix) or email matches dotCom.concat('$') or !(job matches '^'.concat(hasNoJob)) or !(email matches isInvalid.concat('$')))";
-const jsonLogicQueryObject = {
-  and: [
-    { '==': [{ var: 'firstName' }, null] },
-    { '!=': [{ var: 'lastName' }, null] },
-    { in: [{ var: 'firstName' }, ['Test', 'This']] },
-    { '!': { in: [{ var: 'lastName' }, ['Test', 'This']] } },
-    { '<=': ['Test', { var: 'firstName' }, 'This'] },
-    { '<=': ['Test', { var: 'firstName' }, 'This'] },
-    { '!': { '<=': ['Test', { var: 'lastName' }, 'This'] } },
-    { '<=': [12, { var: 'age' }, 14] },
-    { '==': [{ var: 'age' }, '26'] },
-    { '==': [{ var: 'isMusician' }, true] },
-    { '==': [{ var: 'isLucky' }, false] },
-    {
-      '!': {
-        or: [
-          { '==': [{ var: 'gender' }, 'M'] },
-          { '!=': [{ var: 'job' }, 'Programmer'] },
-          { in: ['@', { var: 'email' }] },
-        ],
-      },
-    },
-    {
-      or: [
-        { '!': { in: ['ab', { var: 'lastName' }] } },
-        { startsWith: [{ var: 'job' }, 'Prog'] },
-        { endsWith: [{ var: 'email' }, 'com'] },
-        { '!': { startsWith: [{ var: 'job' }, 'Man'] } },
-        { '!': { endsWith: [{ var: 'email' }, 'fr'] } },
-      ],
-    },
-  ],
-};
-const jsonLogicQueryObjectForValueSourceField = {
-  and: [
-    { '==': [{ var: 'firstName' }, null] },
-    { '!=': [{ var: 'lastName' }, null] },
-    {
-      in: [{ var: 'firstName' }, [{ var: 'middleName' }, { var: 'lastName' }]],
-    },
-    {
-      '!': {
-        in: [{ var: 'lastName' }, [{ var: 'middleName' }, { var: 'lastName' }]],
-      },
-    },
-    {
-      '<=': [{ var: 'middleName' }, { var: 'firstName' }, { var: 'lastName' }],
-    },
-    {
-      '<=': [{ var: 'middleName' }, { var: 'firstName' }, { var: 'lastName' }],
-    },
-    {
-      '!': {
-        '<=': [{ var: 'middleName' }, { var: 'lastName' }, { var: 'lastName' }],
-      },
-    },
-    { '==': [{ var: 'age' }, { var: 'iq' }] },
-    { '==': [{ var: 'isMusician' }, { var: 'isCreative' }] },
-    {
-      '!': {
-        or: [
-          { '==': [{ var: 'gender' }, { var: 'someLetter' }] },
-          { '!=': [{ var: 'job' }, { var: 'isBetweenJobs' }] },
-          { in: [{ var: 'atSign' }, { var: 'email' }] },
-        ],
-      },
-    },
-    {
-      or: [
-        { '!': { in: [{ var: 'firstName' }, { var: 'lastName' }] } },
-        { startsWith: [{ var: 'job' }, { var: 'jobPrefix' }] },
-        { endsWith: [{ var: 'email' }, { var: 'dotCom' }] },
-        { '!': { startsWith: [{ var: 'job' }, { var: 'hasNoJob' }] } },
-        { '!': { endsWith: [{ var: 'email' }, { var: 'isInvalid' }] } },
-      ],
-    },
-  ],
-};
-const elasticSearchQueryObject = {
-  bool: {
-    must: [
-      { bool: { must_not: { exists: { field: 'firstName' } } } },
-      { exists: { field: 'lastName' } },
-      { bool: { should: [{ term: { firstName: 'Test' } }, { term: { firstName: 'This' } }] } },
-      { bool: { must_not: [{ term: { lastName: 'Test' } }, { term: { lastName: 'This' } }] } },
-      { range: { firstName: { gte: 'Test', lte: 'This' } } },
-      { range: { firstName: { gte: 'Test', lte: 'This' } } },
-      { bool: { must_not: { range: { lastName: { gte: 'Test', lte: 'This' } } } } },
-      { range: { age: { gte: 12, lte: 14 } } },
-      { term: { age: '26' } },
-      { term: { isMusician: true } },
-      { term: { isLucky: false } },
-      {
-        bool: {
-          must_not: {
-            bool: {
-              should: [
-                { term: { gender: 'M' } },
-                { bool: { must_not: { term: { job: 'Programmer' } } } },
-                { regexp: { email: { value: '@' } } },
-              ],
-            },
-          },
-        },
-      },
-      {
-        bool: {
-          should: [
-            { bool: { must_not: { regexp: { lastName: { value: 'ab' } } } } },
-            { regexp: { job: { value: '^Prog' } } },
-            { regexp: { email: { value: 'com$' } } },
-            { bool: { must_not: { regexp: { job: { value: '^Man' } } } } },
-            { bool: { must_not: { regexp: { email: { value: 'fr$' } } } } },
-          ],
-        },
-      },
-    ],
-  },
-};
-const elasticSearchQueryObjectForValueSourceField = {
-  bool: {
-    must: [
-      { bool: { must_not: { exists: { field: 'firstName' } } } },
-      { exists: { field: 'lastName' } },
-      {
-        bool: {
-          should: [
-            { bool: { filter: { script: { script: `doc['firstName'] == doc['middleName']` } } } },
-            { bool: { filter: { script: { script: `doc['firstName'] == doc['lastName']` } } } },
-          ],
-        },
-      },
-      {
-        bool: {
-          must_not: [
-            { bool: { filter: { script: { script: `doc['lastName'] == doc['middleName']` } } } },
-            { bool: { filter: { script: { script: `doc['lastName'] == doc['lastName']` } } } },
-          ],
-        },
-      },
-      {
-        bool: {
-          filter: {
-            script: {
-              script: `doc['firstName'] >= doc['middleName'] && doc['firstName'] <= doc['lastName']`,
-            },
-          },
-        },
-      },
-      {
-        bool: {
-          filter: {
-            script: {
-              script: `doc['firstName'] >= doc['middleName'] && doc['firstName'] <= doc['lastName']`,
-            },
-          },
-        },
-      },
-      {
-        bool: {
-          filter: {
-            script: {
-              script: `!(doc['lastName'] >= doc['middleName'] && doc['lastName'] <= doc['lastName'])`,
-            },
-          },
-        },
-      },
-      { bool: { filter: { script: { script: `doc['age'] == doc['iq']` } } } },
-      { bool: { filter: { script: { script: `doc['isMusician'] == doc['isCreative']` } } } },
-      {
-        bool: {
-          must_not: {
-            bool: {
-              should: [
-                { bool: { filter: { script: { script: `doc['gender'] == doc['someLetter']` } } } },
-                { bool: { filter: { script: { script: `doc['job'] != doc['isBetweenJobs']` } } } },
-                {
-                  bool: { filter: { script: { script: `doc['email'].contains(doc['atSign'])` } } },
-                },
-              ],
-            },
-          },
-        },
-      },
-      {
-        bool: {
-          should: [
-            {
-              bool: {
-                filter: { script: { script: `!doc['lastName'].contains(doc['firstName'])` } },
-              },
-            },
-            { bool: { filter: { script: { script: `doc['job'].startsWith(doc['jobPrefix'])` } } } },
-            { bool: { filter: { script: { script: `doc['email'].endsWith(doc['dotCom'])` } } } },
-            { bool: { filter: { script: { script: `!doc['job'].startsWith(doc['hasNoJob'])` } } } },
-            {
-              bool: { filter: { script: { script: `!doc['email'].endsWith(doc['isInvalid'])` } } },
-            },
-          ],
-        },
-      },
-    ],
-  },
-};
+  parameterizedNamedSQLString,
+  parameterizedSQLString,
+  params,
+  params_named,
+  query,
+  queryWithValueSourceField,
+  sqlString,
+  sqlStringForValueSourceField,
+} from './formatQueryTestUtils';
+import { defaultValueProcessor, defaultValueProcessorByRule } from './index';
+import { quoteFieldNamesWithArray } from './utils';
 
 it('formats JSON correctly', () => {
   expect(formatQuery(query)).toBe(JSON.stringify(query, null, 2));
@@ -577,141 +68,6 @@ it('formats parameterized named SQL correctly', () => {
   expect(formatQuery(queryWithValueSourceField, 'parameterized_named')).toEqual({
     sql: sqlStringForValueSourceField,
     params: {},
-  });
-});
-
-it('formats to mongo query correctly', () => {
-  expect(JSON.parse(formatQuery(mongoQuery, 'mongodb'))).toEqual(mongoQueryExpectation);
-  expect(JSON.parse(formatQuery(mongoQueryWithValueSourceField, 'mongodb'))).toEqual(
-    mongoQueryExpectationForValueSourceField
-  );
-  expect(
-    JSON.parse(
-      formatQuery(mongoQueryWithValueSourceField, {
-        format: 'mongodb',
-        valueProcessor: defaultMongoDBValueProcessor,
-      })
-    )
-  ).toEqual(mongoQueryExpectationForValueSourceField);
-});
-
-it('formats CEL correctly', () => {
-  const celQuery = add(query, { field: 'invalid', operator: 'invalid', value: '' }, []);
-  expect(formatQuery(celQuery, 'cel')).toBe(celString);
-  expect(formatQuery(queryWithValueSourceField, 'cel')).toBe(celStringForValueSourceField);
-  expect(
-    formatQuery(queryWithValueSourceField, {
-      format: 'cel',
-      valueProcessor: defaultCELValueProcessor,
-    })
-  ).toBe(celStringForValueSourceField);
-  expect(
-    formatQuery(
-      { combinator: 'and', rules: [{ field: 'f', operator: 'between', value: [14, 12] }] },
-      'cel'
-    )
-  ).toBe('(f >= 12 && f <= 14)');
-});
-
-it('formats SpEL correctly', () => {
-  const spelQuery = add(query, { field: 'invalid', operator: 'invalid', value: '' }, []);
-  expect(formatQuery(spelQuery, 'spel')).toBe(spelString);
-  expect(formatQuery(queryWithValueSourceField, 'spel')).toBe(spelStringForValueSourceField);
-  expect(
-    formatQuery(queryWithValueSourceField, {
-      format: 'spel',
-      valueProcessor: defaultSpELValueProcessor,
-    })
-  ).toBe(spelStringForValueSourceField);
-  expect(
-    formatQuery(
-      { combinator: 'and', rules: [{ field: 'f', operator: 'between', value: [14, 12] }] },
-      'spel'
-    )
-  ).toBe('(f >= 12 and f <= 14)');
-});
-
-it('formats JSONLogic correctly', () => {
-  // Add an empty group and a rule with a non-standard operator
-  const jsonLogicQuery = add(
-    add(query, { combinator: 'and', rules: [] }, []),
-    { field: 'invalid', operator: 'invalid', value: '' },
-    []
-  );
-  expect(formatQuery(jsonLogicQuery, 'jsonlogic')).toEqual(jsonLogicQueryObject);
-  expect(formatQuery(queryWithValueSourceField, 'jsonlogic')).toEqual(
-    jsonLogicQueryObjectForValueSourceField
-  );
-  expect(
-    formatQuery(
-      {
-        combinator: 'and',
-        rules: [{ field: 'f', operator: 'between', value: [14, 12] }],
-      },
-      'jsonlogic'
-    )
-  ).toEqual({ '<=': [12, { var: 'f' }, 14] });
-  expect(
-    formatQuery(
-      {
-        combinator: 'and',
-        rules: [
-          { field: 'f', operator: 'between', value: [14, 12] },
-          {
-            combinator: 'or',
-            rules: [{ field: 'f', operator: '=', value: '26' }],
-          },
-        ],
-      },
-      'jsonlogic'
-    )
-  ).toEqual({ and: [{ '<=': [12, { var: 'f' }, 14] }, { or: [{ '==': [{ var: 'f' }, '26'] }] }] });
-});
-
-it('formats ElasticSearch correctly', () => {
-  expect(formatQuery(query, 'elasticsearch')).toEqual(elasticSearchQueryObject);
-  expect(formatQuery(queryWithValueSourceField, 'elasticsearch')).toEqual(
-    elasticSearchQueryObjectForValueSourceField
-  );
-  expect(
-    formatQuery(
-      {
-        combinator: 'and',
-        rules: [
-          // combinator: 'and', not: true
-          { combinator: 'and', not: true, rules: [{ field: 'f1', operator: '=', value: 'v1' }] },
-          // Weird field names
-          { field: "f\\'1", operator: 'contains', value: 'v1', valueSource: 'field' },
-          // Ranges
-          { field: 'f1', operator: '<', value: 0 },
-          { field: 'f1', operator: '<=', value: 0 },
-          { field: 'f1', operator: '>', value: 0 },
-          { field: 'f1', operator: '>=', value: 0 },
-          { field: 'f1', operator: 'between', value: [10, 1] },
-          // Invalid operator
-          { field: 'f1', operator: 'invalid', value: 'v1' },
-          // Invalid value as field
-          { field: 'f1', operator: 'invalid', value: ['v1', 0], valueSource: 'field' },
-          { field: 'f1', operator: '=', value: '', valueSource: 'field' },
-          { field: 'f1', operator: 'contains', value: '', valueSource: 'field' },
-          // Empty group
-          { combinator: 'or', rules: [] },
-        ],
-      },
-      'elasticsearch'
-    )
-  ).toEqual({
-    bool: {
-      must: [
-        { bool: { must_not: [{ term: { f1: 'v1' } }] } },
-        { bool: { filter: { script: { script: `doc['f\\\\\\'1'].contains(doc['v1'])` } } } },
-        { range: { f1: { lt: 0 } } },
-        { range: { f1: { lte: 0 } } },
-        { range: { f1: { gt: 0 } } },
-        { range: { f1: { gte: 0 } } },
-        { range: { f1: { gte: 1, lte: 10 } } },
-      ],
-    },
   });
 });
 
@@ -880,25 +236,10 @@ describe('escapes quotes when appropriate', () => {
     { fmt: 'sql', result: `(f1 = 'Te''st')` },
     { fmt: 'parameterized', result: { sql: `(f1 = ?)`, params: [`Te'st`] } },
     { fmt: 'parameterized_named', result: { sql: `(f1 = :f1_1)`, params: { f1_1: `Te'st` } } },
-    { fmt: 'spel', result: `f1 == 'Te\\'st'` },
   ]) {
     it(`escapes single quotes (if appropriate) for ${attempt.fmt} export`, () => {
       // @ts-expect-error Conflicting formatQuery overloads
       expect(formatQuery(testQuerySQ, attempt.fmt)).toEqual(attempt.result);
-    });
-  }
-
-  const testQueryDQ: RuleGroupType = {
-    combinator: 'and',
-    rules: [{ field: 'f1', operator: '=', value: `Te"st` }],
-  };
-
-  for (const attempt of [
-    { fmt: 'mongodb', result: `{"f1":"Te\\"st"}` },
-    { fmt: 'cel', result: `f1 == "Te\\"st"` },
-  ]) {
-    it(`escapes double quotes (if appropriate) for ${attempt.fmt} export`, () => {
-      expect(formatQuery(testQueryDQ, attempt.fmt as 'cel' | 'mongodb')).toEqual(attempt.result);
     });
   }
 });
@@ -931,53 +272,6 @@ describe('independent combinators', () => {
     expect(formatQuery(queryIC, 'parameterized_named')).toEqual({
       sql: `(firstName = :firstName_1 and middleName = :middleName_1 or lastName = :lastName_1)`,
       params: { firstName_1: 'Test', middleName_1: 'Test', lastName_1: 'Test' },
-    });
-  });
-
-  it('handles independent combinators for cel', () => {
-    expect(formatQuery(queryIC, 'cel')).toBe(
-      `firstName == "Test" && middleName == "Test" || lastName == "Test"`
-    );
-  });
-
-  it('handles independent combinators for spel', () => {
-    expect(formatQuery(queryIC, 'spel')).toBe(
-      `firstName == 'Test' and middleName == 'Test' or lastName == 'Test'`
-    );
-  });
-
-  it('handles independent combinators for mongodb', () => {
-    expect(formatQuery(queryIC, 'mongodb')).toBe(
-      '{"$or":[{"$and":[{"firstName":"Test"},{"middleName":"Test"}]},{"lastName":"Test"}]}'
-    );
-  });
-
-  it('handles independent combinators for jsonlogic', () => {
-    expect(formatQuery(queryIC, 'jsonlogic')).toEqual({
-      or: [
-        {
-          and: [
-            { '==': [{ var: 'firstName' }, 'Test'] },
-            { '==': [{ var: 'middleName' }, 'Test'] },
-          ],
-        },
-        { '==': [{ var: 'lastName' }, 'Test'] },
-      ],
-    });
-  });
-
-  it('handles independent combinators for elasticsearch', () => {
-    expect(formatQuery(queryIC, 'elasticsearch')).toEqual({
-      bool: {
-        should: [
-          {
-            bool: {
-              must: [{ term: { firstName: 'Test' } }, { term: { middleName: 'Test' } }],
-            },
-          },
-          { term: { lastName: 'Test' } },
-        ],
-      },
     });
   });
 });
@@ -1147,356 +441,6 @@ describe('validation', () => {
       ).toEqual({ sql: '(1 = 1)', params: {} });
     });
   });
-
-  describe('mongodb', () => {
-    it('should invalidate a mongodb query', () => {
-      expect(
-        formatQuery(
-          { id: 'root', combinator: 'and', rules: [] },
-          { format: 'mongodb', validator: () => false }
-        )
-      ).toBe('{"$and":[{"$expr":true}]}');
-    });
-
-    it('should invalidate a mongodb rule', () => {
-      expect(
-        formatQuery(
-          {
-            id: 'root',
-            combinator: 'and',
-            rules: [
-              { field: 'field', operator: '=', value: '' },
-              { field: 'otherfield', operator: '=', value: '' },
-            ],
-          },
-          {
-            format: 'mongodb',
-            fields: [{ name: 'field', label: 'field', validator: () => false }],
-          }
-        )
-      ).toBe('{"otherfield":""}');
-    });
-
-    it('should invalidate mongodb even if fields are valid', () => {
-      expect(
-        formatQuery(
-          {
-            id: 'root',
-            combinator: 'and',
-            rules: [{ field: 'field', operator: '=', value: '' }],
-          },
-          {
-            format: 'mongodb',
-            validator: () => false,
-            fields: [{ name: 'field', label: 'field', validator: () => true }],
-          }
-        )
-      ).toBe('{"$and":[{"$expr":true}]}');
-    });
-
-    it('should invalidate mongodb outermost group', () => {
-      expect(
-        formatQuery(
-          { id: 'root', combinator: 'and', rules: [] },
-          { format: 'mongodb', validator: () => ({ root: false }) }
-        )
-      ).toBe('{"$and":[{"$expr":true}]}');
-    });
-
-    it('should invalidate mongodb inner group', () => {
-      expect(
-        formatQuery(
-          {
-            id: 'root',
-            combinator: 'and',
-            rules: [{ id: 'inner', combinator: 'and', rules: [] }],
-          },
-          {
-            format: 'mongodb',
-            validator: () => ({ inner: false }),
-          }
-        )
-      ).toBe('{"$and":[{"$expr":true}]}');
-    });
-  });
-
-  describe('cel', () => {
-    it('should invalidate a cel query', () => {
-      expect(
-        formatQuery(
-          { id: 'root', combinator: 'and', rules: [] },
-          { format: 'cel', validator: () => false }
-        )
-      ).toBe('1 == 1');
-    });
-
-    it('should invalidate a cel rule', () => {
-      expect(
-        formatQuery(
-          {
-            id: 'root',
-            combinator: 'and',
-            rules: [
-              { field: 'field', operator: '=', value: '' },
-              { field: 'otherfield', operator: '=', value: '' },
-            ],
-          },
-          { format: 'cel', fields: [{ name: 'field', label: 'field', validator: () => false }] }
-        )
-      ).toBe('otherfield == ""');
-    });
-
-    it('should invalidate cel even if fields are valid', () => {
-      expect(
-        formatQuery(
-          {
-            id: 'root',
-            combinator: 'and',
-            rules: [{ field: 'field', operator: '=', value: '' }],
-          },
-          {
-            format: 'cel',
-            validator: () => false,
-            fields: [{ name: 'field', label: 'field', validator: () => true }],
-          }
-        )
-      ).toBe('1 == 1');
-    });
-
-    it('should invalidate cel outermost group', () => {
-      expect(
-        formatQuery(
-          { id: 'root', combinator: 'and', rules: [] },
-          { format: 'cel', validator: () => ({ root: false }) }
-        )
-      ).toBe('1 == 1');
-    });
-
-    it('should invalidate cel inner group', () => {
-      expect(
-        formatQuery(
-          {
-            id: 'root',
-            combinator: 'and',
-            rules: [{ id: 'inner', combinator: 'and', rules: [] }],
-          },
-          {
-            format: 'cel',
-            validator: () => ({ inner: false }),
-          }
-        )
-      ).toBe('1 == 1');
-    });
-  });
-
-  describe('spel', () => {
-    it('should invalidate a spel query', () => {
-      expect(
-        formatQuery(
-          { id: 'root', combinator: 'and', rules: [] },
-          { format: 'spel', validator: () => false }
-        )
-      ).toBe('1 == 1');
-    });
-
-    it('should invalidate a spel rule', () => {
-      expect(
-        formatQuery(
-          {
-            id: 'root',
-            combinator: 'and',
-            rules: [
-              { field: 'field', operator: '=', value: '' },
-              { field: 'otherfield', operator: '=', value: '' },
-            ],
-          },
-          {
-            format: 'spel',
-            fields: [{ name: 'field', label: 'field', validator: () => false }],
-          }
-        )
-      ).toBe("otherfield == ''");
-    });
-
-    it('should invalidate spel even if fields are valid', () => {
-      expect(
-        formatQuery(
-          {
-            id: 'root',
-            combinator: 'and',
-            rules: [{ field: 'field', operator: '=', value: '' }],
-          },
-          {
-            format: 'spel',
-            validator: () => false,
-            fields: [{ name: 'field', label: 'field', validator: () => true }],
-          }
-        )
-      ).toBe('1 == 1');
-    });
-
-    it('should invalidate spel outermost group', () => {
-      expect(
-        formatQuery(
-          { id: 'root', combinator: 'and', rules: [] },
-          { format: 'spel', validator: () => ({ root: false }) }
-        )
-      ).toBe('1 == 1');
-    });
-
-    it('should invalidate spel inner group', () => {
-      expect(
-        formatQuery(
-          {
-            id: 'root',
-            combinator: 'and',
-            rules: [{ id: 'inner', combinator: 'and', rules: [] }],
-          },
-          {
-            format: 'spel',
-            validator: () => ({ inner: false }),
-          }
-        )
-      ).toBe('1 == 1');
-    });
-  });
-
-  describe('jsonlogic', () => {
-    it('should invalidate a jsonlogic query', () => {
-      expect(
-        formatQuery(
-          { id: 'root', combinator: 'and', rules: [] },
-          { format: 'jsonlogic', validator: () => false }
-        )
-      ).toBe(false);
-    });
-
-    it('should invalidate a jsonlogic rule', () => {
-      expect(
-        formatQuery(
-          {
-            id: 'root',
-            combinator: 'and',
-            rules: [
-              { field: 'field', operator: '=', value: '' },
-              { field: 'otherfield', operator: '=', value: '' },
-            ],
-          },
-          {
-            format: 'jsonlogic',
-            fields: [{ name: 'field', label: 'field', validator: () => false }],
-          }
-        )
-      ).toEqual({ '==': [{ var: 'otherfield' }, ''] });
-    });
-
-    it('should invalidate jsonlogic even if fields are valid', () => {
-      expect(
-        formatQuery(
-          {
-            id: 'root',
-            combinator: 'and',
-            rules: [{ field: 'field', operator: '=', value: '' }],
-          },
-          {
-            format: 'jsonlogic',
-            validator: () => false,
-            fields: [{ name: 'field', label: 'field', validator: () => true }],
-          }
-        )
-      ).toBe(false);
-    });
-
-    it('should invalidate jsonlogic outermost group', () => {
-      expect(
-        formatQuery(
-          { id: 'root', combinator: 'and', rules: [] },
-          { format: 'jsonlogic', validator: () => ({ root: false }) }
-        )
-      ).toBe(false);
-    });
-
-    it('should invalidate jsonlogic inner group', () => {
-      expect(
-        formatQuery(
-          {
-            id: 'root',
-            combinator: 'and',
-            rules: [{ id: 'inner', combinator: 'and', rules: [] }],
-          },
-          {
-            format: 'jsonlogic',
-            validator: () => ({ inner: false }),
-          }
-        )
-      ).toBe(false);
-    });
-  });
-
-  describe('elasticsearch', () => {
-    it('should invalidate a elasticsearch query', () => {
-      expect(
-        formatQuery(
-          { id: 'root', combinator: 'and', rules: [] },
-          { format: 'elasticsearch', validator: () => false }
-        )
-      ).toEqual({});
-    });
-
-    it('should invalidate a elasticsearch rule', () => {
-      expect(
-        formatQuery(
-          {
-            id: 'root',
-            combinator: 'and',
-            rules: [
-              { field: 'field', operator: '=', value: '' },
-              { field: 'otherfield', operator: '=', value: '' },
-            ],
-          },
-          {
-            format: 'elasticsearch',
-            fields: [{ name: 'field', label: 'field', validator: () => false }],
-          }
-        )
-      ).toEqual({ bool: { must: [{ term: { otherfield: '' } }] } });
-    });
-
-    it('should invalidate elasticsearch even if fields are valid', () => {
-      expect(
-        formatQuery(
-          {
-            id: 'root',
-            combinator: 'and',
-            rules: [{ field: 'field', operator: '=', value: '' }],
-          },
-          {
-            format: 'elasticsearch',
-            validator: () => false,
-            fields: [{ name: 'field', label: 'field', validator: () => true }],
-          }
-        )
-      ).toEqual({});
-    });
-
-    it('should invalidate elasticsearch outermost group', () => {
-      expect(
-        formatQuery(
-          { id: 'root', combinator: 'and', rules: [] },
-          { format: 'elasticsearch', validator: () => ({ root: false }) }
-        )
-      ).toEqual({});
-    });
-
-    it('should invalidate elasticsearch inner group', () => {
-      expect(
-        formatQuery(
-          { id: 'root', combinator: 'and', rules: [{ id: 'inner', combinator: 'and', rules: [] }] },
-          { format: 'elasticsearch', validator: () => ({ inner: false }) }
-        )
-      ).toEqual({});
-    });
-  });
 });
 
 describe('ruleProcessor', () => {
@@ -1514,68 +458,6 @@ describe('ruleProcessor', () => {
     expect(formatQuery(queryForRuleProcessor, { format: 'sql', ruleProcessor })).toBe(
       "(custom_operator and f2 = 'v2')"
     );
-  });
-
-  it('handles custom MongoDB rule processor', () => {
-    const ruleProcessor: RuleProcessor = r =>
-      r.operator === 'custom_operator' ? r.operator : defaultRuleProcessorMongoDB(r);
-    expect(formatQuery(queryForRuleProcessor, { format: 'mongodb', ruleProcessor })).toBe(
-      '{"$and":[custom_operator,{"f2":"v2"}]}'
-    );
-    expect(
-      formatQuery(queryForRuleProcessor, { format: 'mongodb', valueProcessor: ruleProcessor })
-    ).toBe('{"$and":[custom_operator,{"f2":"v2"}]}');
-  });
-
-  it('handles custom CEL rule processor', () => {
-    const ruleProcessor: RuleProcessor = r =>
-      r.operator === 'custom_operator' ? r.operator : defaultRuleProcessorCEL(r);
-    expect(formatQuery(queryForRuleProcessor, { format: 'cel', ruleProcessor })).toBe(
-      'custom_operator && f2 == "v2"'
-    );
-    expect(
-      formatQuery(queryForRuleProcessor, { format: 'cel', valueProcessor: ruleProcessor })
-    ).toBe('custom_operator && f2 == "v2"');
-  });
-
-  it('handles custom SpEL rule processor', () => {
-    const ruleProcessor: RuleProcessor = r =>
-      r.operator === 'custom_operator' ? r.operator : defaultRuleProcessorSpEL(r);
-    expect(formatQuery(queryForRuleProcessor, { format: 'spel', ruleProcessor })).toBe(
-      "custom_operator and f2 == 'v2'"
-    );
-    expect(
-      formatQuery(queryForRuleProcessor, { format: 'spel', valueProcessor: ruleProcessor })
-    ).toBe("custom_operator and f2 == 'v2'");
-  });
-
-  it('handles custom JsonLogic rule processor', () => {
-    const ruleProcessor: RuleProcessor = r =>
-      r.operator === 'custom_operator'
-        ? { [r.operator]: [{ var: r.field }, r.value] }
-        : defaultRuleProcessorJsonLogic(r);
-    expect(formatQuery(queryForRuleProcessor, { format: 'jsonlogic', ruleProcessor })).toEqual({
-      and: [{ custom_operator: [{ var: 'f1' }, 'v1'] }, { '==': [{ var: 'f2' }, 'v2'] }],
-    });
-    expect(
-      formatQuery(queryForRuleProcessor, { format: 'jsonlogic', valueProcessor: ruleProcessor })
-    ).toEqual({
-      and: [{ custom_operator: [{ var: 'f1' }, 'v1'] }, { '==': [{ var: 'f2' }, 'v2'] }],
-    });
-  });
-
-  it('handles custom ElasticSearch rule processor', () => {
-    const customResult = { bool: { must: [{ term: { custom: 'custom' } }] } };
-    const ruleProcessor: RuleProcessor = r =>
-      r.operator === 'custom_operator' ? customResult : defaultRuleProcessorElasticSearch(r);
-    expect(formatQuery(queryForRuleProcessor, { format: 'elasticsearch', ruleProcessor })).toEqual({
-      bool: { must: [customResult, { term: { f2: 'v2' } }] },
-    });
-    expect(
-      formatQuery(queryForRuleProcessor, { format: 'elasticsearch', valueProcessor: ruleProcessor })
-    ).toEqual({
-      bool: { must: [customResult, { term: { f2: 'v2' } }] },
-    });
   });
 });
 
@@ -1809,101 +691,6 @@ describe('parseNumbers', () => {
       f_23: {},
     });
   });
-
-  it('parses numbers for mongodb', () => {
-    expect(formatQuery(queryForNumberParsing, { format: 'mongodb', parseNumbers: true })).toBe(
-      '{"$and":[{"f":{"$gt":"NaN"}},{"f":0},{"f":0},{"f":0},{"$or":[{"f":{"$lt":1.5}},{"f":{"$gt":1.5}}]},{"f":{"$in":[0,1,2]}},{"f":{"$in":[0,1,2]}},{"f":{"$in":[0,"abc",2]}},{"f":{"$gte":0,"$lte":1}},{"f":{"$gte":0,"$lte":1}},{"f":{"$gte":0,"$lte":"abc"}},{"f":{"$gte":"[object Object]","$lte":"[object Object]"}}]}'
-    );
-  });
-
-  it('parses numbers for cel', () => {
-    expect(formatQuery(queryForNumberParsing, { format: 'cel', parseNumbers: true })).toBe(
-      'f > "NaN" && f == 0 && f == 0 && f == 0 && (f < 1.5 || f > 1.5) && f in [0, 1, 2] && f in [0, 1, 2] && f in [0, "abc", 2] && (f >= 0 && f <= 1) && (f >= 0 && f <= "abc") && (f >= "[object Object]" && f <= "[object Object]")'
-    );
-    const queryForNumberParsingCEL: RuleGroupType = {
-      combinator: 'and',
-      rules: [
-        { field: 'f', operator: 'beginsWith', value: 1 },
-        { field: 'f', operator: 'endsWith', value: 1 },
-      ],
-    };
-    expect(formatQuery(queryForNumberParsingCEL, { format: 'cel', parseNumbers: true })).toBe(
-      `f.startsWith("1") && f.endsWith("1")`
-    );
-  });
-
-  it('parses numbers for spel', () => {
-    expect(formatQuery(queryForNumberParsing, { format: 'spel', parseNumbers: true })).toBe(
-      "f > 'NaN' and f == 0 and f == 0 and f == 0 and (f < 1.5 or f > 1.5) and (f == 0 or f == 1 or f == 2) and (f == 0 or f == 1 or f == 2) and (f == 0 or f == 'abc' or f == 2) and (f >= 0 and f <= 1) and (f >= 0 and f <= 'abc') and (f >= '[object Object]' and f <= '[object Object]')"
-    );
-    const queryForNumberParsingSpEL: RuleGroupType = {
-      combinator: 'and',
-      rules: [
-        { field: 'f', operator: 'beginsWith', value: 1 },
-        { field: 'f', operator: 'beginsWith', value: '^hasCaret' },
-        { field: 'f', operator: 'endsWith', value: 1 },
-        { field: 'f', operator: 'endsWith', value: 'hasDollarSign$' },
-      ],
-    };
-    expect(
-      formatQuery(queryForNumberParsingSpEL, {
-        format: 'spel',
-        parseNumbers: true,
-      })
-    ).toBe(
-      `f matches '^1' and f matches '^hasCaret' and f matches '1$' and f matches 'hasDollarSign$'`
-    );
-  });
-
-  it('parses numbers for jsonlogic', () => {
-    expect(
-      formatQuery(queryForNumberParsing, {
-        format: 'jsonlogic',
-        parseNumbers: true,
-      })
-    ).toEqual({
-      and: [
-        { '>': [{ var: 'f' }, 'NaN'] },
-        { '==': [{ var: 'f' }, 0] },
-        { '==': [{ var: 'f' }, 0] },
-        { '==': [{ var: 'f' }, 0] },
-        { or: [{ '<': [{ var: 'f' }, 1.5] }, { '>': [{ var: 'f' }, 1.5] }] },
-        { in: [{ var: 'f' }, [0, 1, 2]] },
-        { in: [{ var: 'f' }, [0, 1, 2]] },
-        { in: [{ var: 'f' }, [0, 'abc', 2]] },
-        { '<=': [0, { var: 'f' }, 1] },
-        { '<=': [0, { var: 'f' }, 1] },
-        { '<=': ['0', { var: 'f' }, 'abc'] },
-        { '<=': [{}, { var: 'f' }, {}] },
-      ],
-    });
-  });
-
-  it('parses numbers for elasticsearch', () => {
-    expect(
-      formatQuery(queryForNumberParsing, {
-        format: 'elasticsearch',
-        parseNumbers: true,
-      })
-    ).toEqual({
-      bool: {
-        must: [
-          { range: { f: { gt: 'NaN' } } },
-          { term: { f: 0 } },
-          { term: { f: 0 } },
-          { term: { f: 0 } },
-          { bool: { should: [{ range: { f: { lt: 1.5 } } }, { range: { f: { gt: 1.5 } } }] } },
-          { bool: { should: [{ term: { f: 0 } }, { term: { f: 1 } }, { term: { f: 2 } }] } },
-          { bool: { should: [{ term: { f: 0 } }, { term: { f: 1 } }, { term: { f: 2 } }] } },
-          { bool: { should: [{ term: { f: 0 } }, { term: { f: 'abc' } }, { term: { f: 2 } }] } },
-          { range: { f: { gte: 0, lte: 1 } } },
-          { range: { f: { gte: 0, lte: 1 } } },
-          { range: { f: { gte: '0', lte: 'abc' } } },
-          { range: { f: { gte: {}, lte: {} } } },
-        ],
-      },
-    });
-  });
 });
 
 describe('placeholder names', () => {
@@ -1948,23 +735,5 @@ describe('non-standard combinators', () => {
 describe('misc', () => {
   it('quoteFieldNamesWithArray handles null', () => {
     expect(quoteFieldNamesWithArray(null)).toEqual(['', '']);
-  });
-
-  it('runs the jsonLogic additional operators', () => {
-    const { startsWith, endsWith } = jsonLogicAdditionalOperators;
-    expect(startsWith('TestString', 'Test')).toBe(true);
-    // @ts-expect-error null is not valid
-    expect(startsWith(null, 'Test')).toBe(false);
-    // @ts-expect-error [] is not valid
-    expect(startsWith([], 'Test')).toBe(false);
-    // @ts-expect-error {} is not valid
-    expect(startsWith({}, 'Test')).toBe(false);
-    expect(endsWith('TestString', 'String')).toBe(true);
-    // @ts-expect-error null is not valid
-    expect(endsWith(null, 'String')).toBe(false);
-    // @ts-expect-error [] is not valid
-    expect(endsWith([], 'String')).toBe(false);
-    // @ts-expect-error {} is not valid
-    expect(endsWith({}, 'String')).toBe(false);
   });
 });
